Remove unneeded '!'
When first creating these functions, I found it a smart plan to invent my own convention
of using '!' when creating objects.  Inspired from scheme's custom of using this when
signifying mutation.

I think it is a better idea not to use it, and rather have the corresponding make-*
functions as a convention