Fix failing test and decrease epsilon
I had a failing test, so I increased the loose eq? check so that
the tests failed again. Upon inspecting more closely I saw that
it was the test that was wrong.

-0.81381 should be -0.81391

Fixed, tests are passing, so epsilon is increased again.

Also, removed the *var* convention, since I find it noisy