Compute normal at a point on a sphere
This function computes the normal on a sphere. We have to move
point to object space so that we can continue using unit spheres etc.

First we create the object-point by inverting the transform
connected to a sphere. Now we can compute the world-normal.

We find the object-space-normal, then transform it back again to
world space