Completed heroic refactor
It was a hassle to sort out the dependency circles in the way
the project was set up until now. I simplified by moving
things out of their respective folders.

Things should be cleaner now, and it should be easier to
maintain the system as well as the packages.

However, I need to figure out a way to organize the exports