Optimize inversion of matrix
This optimization is borrowed from .Net - System.Numerics. It
uses three helper functions declaimed to inline, to ease
readability of the maths inside.

We know we are dealing with 4x4 matrix, so we don't need to
loop. We can use math only, however, it creates a quite long
function.

The optimization is well worth it - a small test revealed a
performance increase of approx. 40x just from this function.